Conditionally render details
