+++ src/components/MovieDetails/MovieDetails.js
@@ -1,59 +1,57 @@
 import './MovieDetails.css';
-import backIcon from '../../images/back-icon.png'
+import backIcon from '../../images/back-icon.png';
+import { humanizeDate } from '../../helperFunctions';
+import { useEffect, useState } from 'react';
 
-const MovieDetails = ({ chosenMovie, goToHomeView }) => {
 
-  const humanizeDate = (date) => {
-    const inputDate = date.toString();
-    const year = inputDate.slice(0,4);
-    const month = inputDate.slice(5,7);
-    const newDate = inputDate.slice(8);
-    const months = {
-      "01": "January",
-      "02": "February",
-      "03": "March",
-      "04": "April",
-      "05": "May",
-      "06": "June",
-      "07": "July",
-      "08": "August",
-      "09": "September",
-      "10": "October",
-      "11": "November",
-      "12": "December"
-    };
-    return `${months[month]} ${newDate}, ${year}`;
-  }
+const MovieDetails = ({ chosenMovie, goToHomeView, getData }) => {
+  const [dataArrived, setDataArrived] = useState(false);
+  const [details, setDetails] = useState({});
 
-  return (
-    <div className='single-movie-view'>
-      <div className='single-movie-background'> 
-        <img src={chosenMovie.backdrop_path}/>
-      </div>
-      <div className="details-content">
-        <div className='back-icon-container' onClick={() => {goToHomeView(true)}}>
-          <img src={backIcon}/>
+  useEffect(() => {
+    const chosenID = chosenMovie.id
+    getData(`movies/${chosenID}`)
+      .then(data => {
+        const fetchedMovie = data.movie
+        setDetails(fetchedMovie);
+        setDataArrived(true);
+      })
+  }, [])
+
+  const DetailedView = ({details}) => {
+    return (
+      <div className='single-movie-view'>
+        <div className='single-movie-background'> 
+          <img src={chosenMovie.backdrop_path}/>
         </div>
-        <div className='movie-details-poster-container'>
-          <img src={chosenMovie.poster_path} className='details-movie-poster' alt={chosenMovie.title} id={chosenMovie.id}/>
-        </div>
-        <div className='details'>
-          <h2 className='movie-details-title'>
-            {`${chosenMovie.title}`}
-          </h2>
-          <div className='release-date'>
-            Released on {`${humanizeDate(chosenMovie.release_date)}`}
+        <div className="details-content">
+          <div className='back-icon-container' onClick={() => {goToHomeView(true)}}>
+            <img src={backIcon}/>
           </div>
-          <h3 className='rating'>
-            Average Rating: {chosenMovie.average_rating.toFixed(1)} / 10
-          </h3>
-          <p className='movie-description'> 
-          This is placeholder text:
-          Nearly 5,000 years after he was bestowed with the almighty powers of the Egyptian gods—and imprisoned just as quickly—Black Adam is freed from his earthly tomb, ready to unleash his unique form of justice on the modern world.</p>
+          <div className='movie-details-poster-container'>
+            <img src={chosenMovie.poster_path} className='details-movie-poster' alt={chosenMovie.title} id={chosenMovie.id}/>
+          </div>
+          <div className='details'>
+            <h2 className='movie-details-title'>
+              {`${chosenMovie.title}`}
+            </h2>
+            <div className='release-date'>
+              Released on {`${humanizeDate(chosenMovie.release_date)}`}
+            </div>
+            <h3 className='rating'>
+              Average Rating: {chosenMovie.average_rating.toFixed(1)} / 10
+            </h3>
+            <p className='movie-description'> 
+            This is placeholder text:
+            Nearly 5,000 years after he was bestowed with the almighty powers of the Egyptian gods—and imprisoned just as quickly—Black Adam is freed from his earthly tomb, ready to unleash his unique form of justice on the modern world.</p>
+          </div>
         </div>
       </div>
-    </div>
-  )
+    )
+  }
+  return (<>
+    {dataArrived && <DetailedView />}
+  </>)
 }
 
 export default MovieDetails
